Update saveEXT.py
Changed find_all_dats() function to only return DATs that have a 'file' parameter. I noticed it was coloring parameter, sort, and substitute DATs as external because they do not have a 'file' parameter. I changed it to check if that parameter exists as well as that it's not an empty string before appending it to the external_dats list.